feat: Moves process_host_module to util.
