+++ resources/prosody-plugins/mod_muc_breakout_rooms.lua
@@ -33,6 +33,7 @@ local uuid_gen = require 'util.uuid'.generate;
 local util = module:require 'util';
 local internal_room_jid_match_rewrite = util.internal_room_jid_match_rewrite;
 local is_healthcheck_room = util.is_healthcheck_room;
+local process_host_module = util.process_host_module;
 
 local BREAKOUT_ROOMS_IDENTITY_TYPE = 'breakout_rooms';
 -- only send at most this often updates on breakout rooms to avoid flooding.
@@ -430,25 +431,6 @@ end
 
 -- Module operations
 
--- process a host module directly if loaded or hooks to wait for its load
-function process_host_module(name, callback)
-    local function process_host(host)
-        if host == name then
-            callback(module:context(host), host);
-        end
-    end
-
-    if prosody.hosts[name] == nil then
-        module:log('debug', 'No host/component found, will wait for it: %s', name)
-
-        -- when a host or component is added
-        prosody.events.add_handler('host-activated', process_host);
-    else
-        process_host(name);
-    end
-end
-
-
 -- operates on already loaded breakout rooms muc module
 function process_breakout_rooms_muc_loaded(breakout_rooms_muc, host_module)
     module:log('debug', 'Breakout rooms muc loaded');
@@ -542,7 +524,8 @@ function process_breakout_rooms_muc_loaded(breakout_rooms_muc, host_module)
 end
 
 -- process or waits to process the breakout rooms muc component
-process_host_module(breakout_rooms_muc_component_config, function(host_module, host)
+process_host_module(breakout_rooms_muc_component_config, function(host)
+    local host_module = module:context(host);
     module:log('info', 'Breakout rooms component created %s', host);
 
     local muc_module = prosody.hosts[host].modules.muc;
@@ -571,7 +554,8 @@ function process_main_muc_loaded(main_muc, host_module)
 end
 
 -- process or waits to process the main muc component
-process_host_module(main_muc_component_config, function(host_module, host)
+process_host_module(main_muc_component_config, function(host)
+    local host_module = module:context(host);
     local muc_module = prosody.hosts[host].modules.muc;
 
     if muc_module then
